move scaleEigenvector to external include (reused by gen_eigensolver)
